Setup Basic Grab Test
Set up a scene to test grab interactables and sockets. These will be used for the clue. Next I will import Spencer's custom interactable script and see if it will break the scene. The XR rig in the SCiL branch doesn't recognize the controllers, even in the demo scene. I'm thinking it had to do with adding the custom script.